DC-1757: fixing IT tests
